feat: adding repository interfaces

+++ src/main/java/io/github/isadorabello/arquiteturaspring/livraria/model/Livro.java
@@ -3,6 +3,7 @@ package io.github.isadorabello.arquiteturaspring.livraria.model;
 import jakarta.persistence.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -29,8 +30,8 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale = 2)
-    private Double preco;
-    // private BigDecimal preço;
+    private BigDecimal preco;
+    // private Double preço;
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
